Reorganise VDI classes and mark their APIs as @Experimental, as they might still change
