+++ src/main/resources/graphics/scenery/volumes/vdi/AccumulateVDI.comp
