+++ src/main/resources/graphics/scenery/volumes/vdi/GridCellsToZero.comp
