+++ src/main/resources/graphics/scenery/volumes/vdi/VDIGenerator.comp
