+++ src/test/kotlin/graphics/scenery/tests/examples/volumes/VolumeManagerSwitchingExample.kt
@@ -3,6 +3,7 @@ package graphics.scenery.tests.examples.volumes
 import graphics.scenery.*
 import graphics.scenery.backends.Renderer
 import graphics.scenery.volumes.*
+import graphics.scenery.volumes.vdi.VDIVolumeManager
 import org.joml.Vector3f
 import java.nio.file.Paths
 import kotlin.concurrent.thread
